the day commit

+++ admin/include/view_allpost.php
@@ -21,9 +21,32 @@ switch($bulk_options){
             // if(!$delete_query_sending){
             //     die("QUERY FAILED".mysqli_error($connection))
             // }
-            confirmQuery( $delete_query_sending);
+            confirmQuery($delete_query_sending);
             break;
 
+
+            case 'clone':
+
+    $query="SELECT*FROM posts WHERE post_id={$post_value_id}";
+    $select_posts=mysqli_query($connection,$query);
+    while($row=mysqli_fetch_assoc($select_posts)){
+    $post_id=$row['post_id'];
+    $post_author=$row['post_author'];
+    $post_title=$row['post_title'];
+    $post_category_id=$row['post_category_id'];
+    $post_status=$row['post_status'];
+    $post_images=$row['post_image'];
+    $post_tags=$row['post_tags'];
+    $post_comment=$row['post_comment_count'];
+    $post_date=$row['post_date'];
+    $post_content=$row['post_content'];
+
+    }
+    $query="INSERT INTO posts(post_category_id,post_title,post_author,post_date,post_image,post_content,post_tags,post_status )";
+$query.="VALUES ({$post_category_id},'{$post_title}','{$post_author}',now(),'{$post_images}','{$post_content}','{ $post_tags}','{$post_status}')";
+$copy_post_query=mysqli_query($connection,$query);
+confirmQuery($copy_post_query);
+break;
   }
 }
 }
@@ -38,6 +61,7 @@ switch($bulk_options){
 <option value="published">Publish</option>
 <option value="drafted">Draft</option>   
 <option value="delete">Delete</option>
+<option value="clone">Clone</option>
 </select>
 </div>
 
@@ -62,11 +86,12 @@ switch($bulk_options){
         <th>View</th>
         <th>EDIT</th>
         <th>DELETE</th>
+        <th>Viewed</th>
         </tr>
 </thead>
 <tbody>
 <?php
-$query="SELECT*FROM posts";
+$query="SELECT*FROM posts ORDER BY post_id DESC ";
 $select_posts=mysqli_query($connection,$query);
 while($row=mysqli_fetch_assoc($select_posts)){
 $post_id=$row['post_id'];
@@ -79,6 +104,7 @@ $post_tags=$row['post_tags'];
 $post_comment=$row['post_comment_count'];
 $post_date=$row['post_date'];
 $post_content=$row['post_content'];
+$post_views_count=$row['post_views_count'];
 
 echo"<tr>";
 ?>
@@ -114,8 +140,10 @@ echo "<td>{$post_date}</td>";
 echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
 echo "
 <td><a  href='posts.php?source=edit_post&p_id={$post_id}'>EDIT</a></td>";
-echo "<td><a href='posts.php?delete={$post_id}'>DELETE</a></td>";
+echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete');\" href='posts.php?delete={$post_id}'>DELETE</a></td>";
+echo "<td><a href='posts.php?reset={$post_id}'>{$post_views_count}</a></td>";
 echo "</tr>";
+
 }
 ?>  
 </tbody>
@@ -128,4 +156,15 @@ $query = "DELETE FROM posts WHERE post_id = {$the_post_id}";
 $delete_query=mysqli_query($connection,$query);
 header("location:posts.php");
 }
+
+if(isset($_GET['reset'])){
+    $the_reset_id=$_GET['reset'];
+    $reset_query ="UPDATE posts SET post_views_count=0 WHERE post_id=".mysqli_real_escape_string($connection,$_GET['reset'])."";
+    $reset_query=mysqli_query($connection,$reset_query);
+
+    if(!$reset_query){
+    die("QUERY FAILED".mysqli_error($connection));
+    }
+    header("location:posts.php");
+    }
 ?>
